projects/toil-automation-identifier/requirements.txt
# Core dependencies
click>=8.1.0
rich>=13.7.0
gitpython>=3.1.40
python-dotenv>=1.0.0

# AI providers (optional)
openai>=1.12.0
ollama>=0.1.0

# Data processing
pandas>=2.2.0
